Use Phil Johnston's headless-blocks npm package for blocks
Render blocks using <HeadlessBlocks>,
which will parse the post.content.
When that finds block comments like
< wp:genesis-custom-blocks/example-block {"title":"Here is some text: "} /-->,
then int looks to see if it was passed
a block in blockComponents.
If so, it renders that component
with the attributes.
For other blocks like Core blocks,
it just renders the static HTML that it's passed.