🔨 refactor(previews): `usePreviewNode` interface
Rename the returned property `postType` of `usePreviewNode` to `typeName` to
better reflect that the property is the GraphQL Type name, and not the WordPress
post type name.

Additionally included `typeName` in the preview URL from the WordPress side
to reduce the need to make a second request for the preview node form Faust.js.

Related issue: #619