test: confirm post_link filtering behavior.
Confirms links are not rewritten on post-new.php pages when $_POST is empty. Also confirms Ajax requests to generate sample permalinks are not rewritten. These checks prevent the incorrect guid from being written to the database.

Confirms `wp-link-ajax` Ajax requests are not rewritten. This is used by the Classic Editor when linking content.

chore: update `enable_rewrites` values in tests to be `1` instead of `true`, for consistency.