MERL-1270: Do not rewrite links in the post editor (#1625)
* fix: do not rewrite links in the post editor

When Faust’s “Enable Post and Category URL rewrites” is enabled, this has been causing URLs to be rewritten in the editor to contain the front-end URL instead of the WP URL, which causes them to be saved to the database this way. Instead, we want the WP URL saved to the database.

* fix: don't rewrite links in `wp-link-ajax` ajax requests

* fix: do not rewrite links on permalink generation requests

* test: confirm post_link filtering behavior.

Confirms links are not rewritten on post-new.php pages when $_POST is empty. Also confirms Ajax requests to generate sample permalinks are not rewritten. These checks prevent the incorrect guid from being written to the database.

Confirms `wp-link-ajax` Ajax requests are not rewritten. This is used by the Classic Editor when linking content.

chore: update `enable_rewrites` values in tests to be `1` instead of `true`, for consistency.

* add changeset

* refactor: extract conditional logic to functions