chore: Update .nvmrc to allow v18 upgrades
Pinning to 18.0.0 causes automated tools like `fnm` to try installing 18.0.0, instead of using the latest 18.x already installed.

For clarity, I have these versions of Node installed:
```
fnm list
* v16.13.0
* v16.15.0
* v16.20.2
* v18.16.1 default
* v18.18.2
* v18.19.0
* v20.8.1
* system
```

I have fnm configured to auto install/switch to the version specified in `.nvmrc`. With the existing config, it says it can't find 18.0.0 and asks if I want to install it. With this updated config, it switches to the latest 18.x that I already have.